+++ packages/terra-core-docs/CHANGELOG.md
@@ -4,10 +4,15 @@
 
 * Added
   * Added accessibility guide for `terra-hyperlink`
+  * Added Accessibility guideline for `terra-heading`
+  * Added examples for for `terra-list`
   * Added Accessibility Guide for `terra-section-header`
   * Added accessibility guide for `terra-progress-bar`
-
+  * Added accessibility guide for `terra-avatar`
+  
 * Changed
+  * Added tests for `terra-list`.
+  * Fix second example on default Paginator doc page
   * Updated example for `terra-hyperlink`
   * Updated example for `terra-progress-bar`